abstract replicator injector and introduce SqliteInjector
