+++ libsql-replication/src/injector/error.rs
@@ -1,3 +1,5 @@
+pub type Result<T, E=Error> = std::result::Result<T, E>;
+
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error("IO error: {0}")]
