+++ libsql-replication/src/replicator.rs
@@ -1,14 +1,11 @@
 use std::path::PathBuf;
-use std::sync::Arc;
 
-use parking_lot::Mutex;
-use tokio::task::spawn_blocking;
 use tokio::time::Duration;
 use tokio_stream::{Stream, StreamExt};
 use tonic::{Code, Status};
 
 use crate::frame::{Frame, FrameNo};
-use crate::injector::Injector;
+use crate::injector::{Injector, SqliteInjector};
 use crate::rpc::replication::{
     Frame as RpcFrame, NAMESPACE_DOESNT_EXIST, NEED_SNAPSHOT_ERROR_MSG, NO_HELLO_ERROR_MSG,
 };
@@ -137,9 +134,9 @@ where
 
 /// The `Replicator`'s duty is to download frames from the primary, and pass them to the injector at
 /// transaction boundaries.
-pub struct Replicator<C> {
+pub struct Replicator<C, I> {
     client: C,
-    injector: Arc<Mutex<Injector>>,
+    injector: I,
     state: ReplicatorState,
     frames_synced: usize,
 }
@@ -154,33 +151,42 @@ enum ReplicatorState {
     Exit,
 }
 
-impl<C: ReplicatorClient> Replicator<C> {
+impl<C> Replicator<C, SqliteInjector>
+where
+    C: ReplicatorClient,
+{
     /// Creates a replicator for the db file pointed at by `db_path`
-    pub async fn new(
+    pub async fn new_sqlite(
         client: C,
         db_path: PathBuf,
         auto_checkpoint: u32,
         encryption_config: Option<libsql_sys::EncryptionConfig>,
     ) -> Result<Self, Error> {
-        let injector = {
-            let db_path = db_path.clone();
-            spawn_blocking(move || {
-                Injector::new(
-                    db_path,
-                    INJECTOR_BUFFER_CAPACITY,
-                    auto_checkpoint,
-                    encryption_config,
-                )
-            })
-            .await??
-        };
+        let injector = SqliteInjector::new(
+            db_path.clone(),
+            INJECTOR_BUFFER_CAPACITY,
+            auto_checkpoint,
+            encryption_config,
+        )
+            .await?;
 
-        Ok(Self {
+        Ok(Self::new(client, injector))
+    }
+}
+
+impl<C, I> Replicator<C, I>
+where
+    C: ReplicatorClient,
+    I: Injector,
+{
+
+    pub fn new(client: C, injector: I) -> Self {
+        Self {
             client,
-            injector: Arc::new(Mutex::new(injector)),
+            injector,
             state: ReplicatorState::NeedHandshake,
             frames_synced: 0,
-        })
+        }
     }
 
     /// for a handshake on next call to replicate.
@@ -250,7 +256,7 @@ impl<C: ReplicatorClient> Replicator<C> {
         // in case of error we rollback the current injector transaction, and start over.
         if ret.is_err() {
             self.client.rollback();
-            self.injector.lock().rollback();
+            self.injector.rollback().await;
         }
 
         self.state = match ret {
@@ -293,7 +299,8 @@ impl<C: ReplicatorClient> Replicator<C> {
     }
 
     async fn load_snapshot(&mut self) -> Result<(), Error> {
-        self.injector.lock().clear_buffer();
+        self.client.rollback();
+        self.injector.rollback().await;
         loop {
             match self.client.snapshot().await {
                 Ok(mut stream) => {
@@ -315,26 +322,22 @@ impl<C: ReplicatorClient> Replicator<C> {
     async fn inject_frame(&mut self, frame: Frame) -> Result<(), Error> {
         self.frames_synced += 1;
 
-        let injector = self.injector.clone();
-        match spawn_blocking(move || injector.lock().inject_frame(frame)).await? {
-            Ok(Some(commit_fno)) => {
+        match self.injector.inject_frame(frame).await? {
+            Some(commit_fno) => {
                 self.client.commit_frame_no(commit_fno).await?;
             }
-            Ok(None) => (),
-            Err(e) => Err(e)?,
+            None => (),
         }
 
         Ok(())
     }
 
     pub async fn flush(&mut self) -> Result<(), Error> {
-        let injector = self.injector.clone();
-        match spawn_blocking(move || injector.lock().flush()).await? {
-            Ok(Some(commit_fno)) => {
+        match self.injector.flush().await? {
+            Some(commit_fno) => {
                 self.client.commit_frame_no(commit_fno).await?;
             }
-            Ok(None) => (),
-            Err(e) => Err(e)?,
+            None => (),
         }
 
         Ok(())
@@ -395,7 +398,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
 
@@ -438,7 +441,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         // we assume that we already received the handshake and the handshake is not valid anymore
@@ -482,7 +485,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         // we assume that we already received the handshake and the handshake is not valid anymore
@@ -526,7 +529,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         // we assume that we already received the handshake and the handshake is not valid anymore
@@ -568,7 +571,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         // we assume that we already received the handshake and the handshake is not valid anymore
@@ -610,7 +613,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         replicator.state = ReplicatorState::NeedSnapshot;
@@ -653,7 +656,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         // we assume that we already received the handshake and the handshake is not valid anymore
@@ -696,7 +699,7 @@ mod test {
             fn rollback(&mut self) {}
         }
 
-        let mut replicator = Replicator::new(Client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(Client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
         replicator.state = ReplicatorState::NeedHandshake;
@@ -784,7 +787,7 @@ mod test {
             committed_frame_no: None,
         };
 
-        let mut replicator = Replicator::new(client, tmp.path().to_path_buf(), 10000, None)
+        let mut replicator = Replicator::new_sqlite(client, tmp.path().to_path_buf(), 10000, None)
             .await
             .unwrap();
 
@@ -795,7 +798,7 @@ mod test {
             replicator.try_replicate_step().await.unwrap_err(),
             Error::Client(_)
         ));
-        assert!(!replicator.injector.lock().is_txn());
+        assert!(!replicator.injector.inner.lock().is_txn());
         assert!(replicator.client_mut().committed_frame_no.is_none());
         assert_eq!(replicator.state, ReplicatorState::NeedHandshake);
 
@@ -805,7 +808,7 @@ mod test {
         replicator.client_mut().should_error = false;
 
         replicator.try_replicate_step().await.unwrap();
-        assert!(!replicator.injector.lock().is_txn());
+        assert!(!replicator.injector.inner.lock().is_txn());
         assert_eq!(replicator.state, ReplicatorState::Exit);
         assert_eq!(replicator.client_mut().committed_frame_no, Some(6));
     }
